+++ NEWS.md
@@ -11,6 +11,12 @@
 
 * Improved support for models of class `geeglm`.
 
+## Changes to functions
+
+* `check_model()` gains a `show_dots` argument, so show or hide data points. 
+  This is particular useful for models with many observations, where generating
+  the plot would be very slow.
+
 ## Bug Fixes
 
 * Fixes wrong column names in `model_performance()` output for `kmeans` objects